requirements.txt
Pillow==9.4.0
rembg==2.0.30
